+++ Addon/ElvUI_mMediaTag/core/changelog/3.64.lua
@@ -1,0 +1,14 @@
+mMT.Changelog[364]  = {
+	DATE = "03.11.2024",
+	UPDATE = {
+        "[Details-Embedded]: Chatframe texture.",
+        "[System]: Boss IDs.",
+        "[System]: TOC Update"
+     },
+	FIX = {
+        "[Details-Embedded]: Fixed Strata and Frame Level.",
+        "Bug where the Border is not adjust correctly.",
+        "[Minimap-Aspectration]: Fixed Icon positions.",
+        "[System]: Bug in Cata wich prevents to open the Game menu from the mMT Dock & Datatexts."
+     },
+}