+++ Addon/ElvUI_mMediaTag/modules/Dock/mFPSMS.lua
@@ -135,27 +135,11 @@ local function OnClick(self, button)
 		mMT:Dock_Click(self, Config)
 		if button == "LeftButton" then
 			if not _G.GameMenuFrame:IsShown() then
-				if not E.Retail then
-					if _G.VideoOptionsFrame:IsShown() then
-						_G.VideoOptionsFrameCancel:Click()
-					elseif _G.AudioOptionsFrame:IsShown() then
-						_G.AudioOptionsFrameCancel:Click()
-					elseif _G.InterfaceOptionsFrame:IsShown() then
-						_G.InterfaceOptionsFrameCancel:Click()
-					end
-				end
-
 				CloseMenus()
 				CloseAllWindows()
 				ShowUIPanel(_G.GameMenuFrame)
 			else
 				HideUIPanel(_G.GameMenuFrame)
-
-				if E.Retail then
-					MainMenuMicroButton:SetButtonState("NORMAL")
-				else
-					MainMenuMicroButton_SetNormal()
-				end
 			end
 		else
 			E:ToggleOptions()