+++ Addon/ElvUI_mMediaTag/modules/Dock/mVolume.lua
@@ -212,11 +212,7 @@ local function OnClick(self, button)
 		mMT:Dock_Click(self, Config)
 		if button == "LeftButton" then
 			if IsShiftKeyDown() then
-				if E.Retail then
 					_G.Settings.OpenToCategory(_G.Settings.AUDIO_CATEGORY_ID)
-				else
-					ShowOptionsPanel(_G.VideoOptionsFrame, _G.GameMenuFrame, SOUND)
-				end
 				return
 			end