+++ Addon/ElvUI_mMediaTag/modules/datatexts/gamemenu.lua
@@ -230,27 +230,11 @@ local function BuildMenu()
 		icon = AddIcon("gears"),
 		func = function()
 			if not _G.GameMenuFrame:IsShown() then
-				if not E.Retail then
-					if _G.VideoOptionsFrame:IsShown() then
-						_G.VideoOptionsFrameCancel:Click()
-					elseif _G.AudioOptionsFrame:IsShown() then
-						_G.AudioOptionsFrameCancel:Click()
-					elseif _G.InterfaceOptionsFrame:IsShown() then
-						_G.InterfaceOptionsFrameCancel:Click()
-					end
-				end
-
 				CloseMenus()
 				CloseAllWindows()
 				ShowUIPanel(_G.GameMenuFrame)
 			else
 				HideUIPanel(_G.GameMenuFrame)
-
-				if E.Retail then
-					MainMenuMicroButton:SetButtonState("NORMAL")
-				else
-					MainMenuMicroButton_SetNormal()
-				end
 			end
 		end,
 	})